refactor(bundle-size): Removed the use of external MUI
