+++ packages/.eslintrc
@@ -58,6 +58,7 @@
     "@typescript-eslint/no-explicit-any": "error",
     "react-hooks/rules-of-hooks": "error", // Checks rules of Hooks
     "react-hooks/exhaustive-deps": "warn", // Checks effect dependencies
+    "react/jsx-props-no-spreading": "off",
     "no-restricted-imports": [
       "error",
       {
