+++ packages/geoview-core/public/templates/loading-packages.html
@@ -222,18 +222,6 @@
                   'id': 'lakes'
                 }
               ]
-            },
-            {
-              'id': 'xyzTilesLYR8',
-              'name': {
-                'en': 'CARTO Raster Basemap',
-                'fr': 'CARTO Raster Basemap'
-              },
-              'url': {
-                'en': 'https://{s}.basemaps.cartocdn.com/rastertiles/light_all/{z}/{x}/{y}.png',
-                'fr': 'https://{s}.basemaps.cartocdn.com/rastertiles/light_all/{z}/{x}/{y}.png'
-              },
-              'layerType': 'xyzTiles'
             }
           ]
         },
