+++ packages/geoview-core/src/ui/index.ts
@@ -1,6 +1,8 @@
+export * from './autocomplete/autocomplete';
 export * from './button';
 export * from './button-group';
 export * from './circular-progress/circular-progress';
+export * from './checkbox/checkbox';
 export * from './divider/divider';
 export * from './drawer/drawer';
 export * from './fade/fade';
@@ -10,6 +12,9 @@ export * from './list';
 export * from './modal';
 export * from './panel';
 export * from './select/select';
+export * from './slider/slider';
 export * from './snackbar/snackbar';
 export * from './stepper/stepper';
 export * from './text-field/text-field';
+export * from './tooltip/tooltip';
+export * from './typography/typography';
