+++ packages/geoview-core/src/ui/list/checkbox-list/checkbox-list.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 
-import { Typography } from '@mui/material';
 import createStyles from '@mui/styles/createStyles';
 import makeStyles from '@mui/styles/makeStyles';
+import Typography from '@mui/material/Typography';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
