+++ packages/geoview-core/src/ui/panel/default-panels.tsx
@@ -1,4 +1,4 @@
-import { Typography } from '@mui/material';
+import Typography from '@mui/material/Typography';
 import HelpIcon from '@mui/icons-material/Help';
 
 import { TypeButtonPanelProps } from '../../core/types/cgpv-types';
