+++ packages/geoview-core/src/ui/select/select.tsx
@@ -1,132 +1,39 @@
-/* eslint-disable array-callback-return */
-/* eslint-disable react/jsx-props-no-spreading */
-/* eslint-disable @typescript-eslint/no-shadow */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable no-unused-expressions */
-import React, { useState } from 'react';
+import React from 'react';
 
-import {
-  FormControl,
-  FormHelperText,
-  InputLabel,
-  ListSubheader,
-  MenuItem,
-  Select as MaterialSelect,
-  SelectChangeEvent,
-} from '@mui/material';
-import makeStyles from '@mui/styles/makeStyles';
+import { MenuItemProps } from '@mui/material';
+
+import MaterialSelect from '@mui/material/Select';
+import FormControl from '@mui/material/FormControl';
+import InputLabel from '@mui/material/InputLabel';
+import MenuItem from '@mui/material/MenuItem';
 
 import { TypeSelectProps } from '../../core/types/cgpv-types';
 
-const useStyles = makeStyles((theme) => ({
-  formControl: {
-    width: '50%',
-    margin: '15px 0',
-    '& .MuiFormLabel-root.Mui-focused': {
-      color: theme.palette.primary.contrastText,
-      background: theme.palette.primary.light,
-    },
-    '& .MuiOutlinedInput-root.Mui-focused': {
-      border: `1px solid ${theme.palette.primary.contrastText}`,
-    },
-  },
-  label: {
-    position: 'absolute',
-    left: 0,
-    top: 0,
-    transform: 'translate(14px, -9px) scale(0.75)',
-    background: theme.palette.primary.light,
-  },
-  select: {
-    width: '100%',
-  },
-}));
+import { generateId } from '../../core/utils/utilities';
 
 /**
- * Create a customizable Material UI Select
+ * Create a Material UI Select component
  *
- * @param {TypeSelectProps} props the properties passed to the Select component
- * @returns {JSX.Element} the created Select element
+ * @param {TypeSelectProps} props custom select properties
+ * @returns {JSX.Element} the auto complete ui component
  */
 export function Select(props: TypeSelectProps): JSX.Element {
-  const classes = useStyles();
-  const [value, setValue] = useState('');
-  const [multipleValue, setMultipleValue] = useState([]);
-  const { className, style, id, label, selectItems, callBack, helperText, multiple, ...otherProps } = props;
-
-  /**
-   * Runs when a selection is changed
-   *
-   * @param event the selection event
-   */
-  const changeHandler = (event: SelectChangeEvent<string>) => {
-    if (!multiple) setValue(event.target.value);
-    if (multiple) {
-      const {
-        target: { value: targetValue },
-      } = event;
-      setMultipleValue((typeof targetValue === 'string' ? targetValue.split(',') : targetValue) as React.SetStateAction<never[]>);
-    }
-  };
-
-  if (!multiple && typeof callBack === 'function') {
-    callBack(value);
-  } else if (multiple && typeof callBack === 'function') {
-    callBack(multipleValue);
-  }
-
-  const isGrouped = selectItems.some((item: any) => item.category);
-
-  const isDefault = !isGrouped
-    ? selectItems.some((item: any) => item.default)
-    : selectItems.some((item: any) => item.items.some((item: any) => item.default));
-
-  isGrouped &&
-    selectItems.forEach((item: any) => {
-      item.items.forEach((item: any) => {
-        if (value) return;
-        if (item.default) setValue(item.value);
-      });
-    });
-
-  !isGrouped &&
-    selectItems.forEach((item: any) => {
-      if (value) return;
-      if (item.default) setValue(item.value);
-    });
+  const { fullWidth, inputLabel, menuItems, ...selectProps } = props;
 
   return (
-    <FormControl className={classes.formControl} {...otherProps}>
-      <InputLabel className={(isDefault && classes.label) as string} id={id}>
-        {label}
-      </InputLabel>
-      <MaterialSelect
-        className={`${classes.select} ${className && className}`}
-        style={style}
-        labelId={id}
-        id={`select-${id}`}
-        label={label || undefined}
-        value={(!multiple ? value : multipleValue) as string}
-        onChange={changeHandler}
-        multiple={multiple || false}
-        displayEmpty
-      >
-        {isGrouped
-          ? selectItems.map((item: any) => {
-              const options: JSX.Element[] = [];
-              if (item.category) options.push(<ListSubheader>{item.category ? item.category : 'Others'}</ListSubheader>);
-              item.items.map((item: any) => {
-                options.push(<MenuItem value={item.value}>{item.value}</MenuItem>);
-              });
-              return options;
-            })
-          : selectItems.map((item: any) => (
-              <MenuItem key={item.id} value={item.value}>
-                {item.value}
-              </MenuItem>
-            ))}
+    <FormControl fullWidth={fullWidth}>
+      <InputLabel {...inputLabel} />
+      <MaterialSelect {...selectProps}>
+        {menuItems.map((menuItem: MenuItemProps | null) => {
+          if (menuItem) {
+            const menuId = generateId(menuItem.id);
+
+            return <MenuItem key={menuId} {...menuItem} />;
+          }
+
+          return null;
+        })}
       </MaterialSelect>
-      {helperText && <FormHelperText>{helperText}</FormHelperText>}
     </FormControl>
   );
 }
