+++ packages/geoview-core/src/ui/snackbar/snackbar.tsx
@@ -2,7 +2,7 @@ import { useEffect, useContext } from 'react';
 
 import { useTranslation } from 'react-i18next';
 
-import { Button } from '@mui/material';
+import Button from '@mui/material/Button';
 
 import { useSnackbar } from 'notistack';
